+++ demo/src/js/actions.js
@@ -4,3 +4,17 @@ export function updateHtml(html) {
     html
   };
 }
+
+export function updateSelectedExample(example) {
+  return {
+    type: 'UPDATE_SELECTED_EXAMPLE',
+    example
+  };
+}
+
+export function setView(view) {
+  return {
+    type: 'SET_VIEW',
+    view
+  };
+}