+++ src/index.js
@@ -1,3 +1,5 @@
 import HtmlParser from './HtmlParser';
-
 export default HtmlParser;
+
+export { default as processNodes } from './processNodes';
+export { default as convertNodeToElement } from './convertNodeToElement';